adding a few functions to the process_mcs_utils.py file
