+++ asf_core_data/pipeline/mcs/process/process_mcs_utils.py
@@ -10,10 +10,7 @@ import requests
 import time
 import re
 import numpy as np
-
-from asf_core_data.getters.data_getters import load_s3_data
-
-from asf_core_data.config import base_config
+import random
 
 # %%
 
@@ -92,19 +89,42 @@ mcs_companies_dict = {
 }
 
 
-def clean_company_name(company_name):
-    """cleans installation company name by:
-        - removing "company" stopwords like ltd, limited;
-        - lowers company name
-        - removes punctuation
-    Args:
-        company_name (str): Name of company to clean.
-    Returns:
-        company_name (str): Cleaned company name.
+def rename_columns(cols: list) -> list:
     """
-    company_stopwords = ["ltd", "ltd.", "limited", "limited.", "old", "account", "t/a"]
+    Renames a list of strings, e.g. a list of column names, by:
+    - applying lower case;
+    - replacing "heat pump" by "hp";
+    - replacing " " and "/" by "_";
+    - correcting the word address.
 
-    company_name = company_name.lower().translate(str.maketrans("", "", "()-"))
+    Args:
+        cols: original column names
+    Returns:
+        Renamed column names
+    """
+    cols = [c.lower() for c in cols]
+    changes_dict = {"heat pump": "hp", " ": "_", "/": "_", "adddress": "address"}
+    for item in changes_dict.keys():
+        cols = [c.replace(item, changes_dict[item]) for c in cols]
+
+    return cols
+
+
+def clean_company_name(company_name: str) -> str:
+    """
+    Cleans installation company name by:
+        - changing company name to lower case;
+        - removing punctuation;
+        - removing company-related stopwords like "ltd", "limited", etc.;
+
+    Args:
+        company_name: Name of company to clean.
+    Returns:
+        company_name: Cleaned company name.
+    """
+    company_stopwords = ["ltd", "limited", "old", "account"]
+
+    company_name = company_name.lower().translate(str.maketrans("", "", "()-,/.&"))
     company_name = [
         comp for comp in company_name.split() if comp not in company_stopwords
     ]
@@ -112,8 +132,177 @@ def clean_company_name(company_name):
     return " ".join(company_name)
 
 
+def deal_with_versions_of_trading_as(installer_name: str) -> str:
+    """
+    Replaces several expressions representing "trading as" by the expression
+    "trading as" in the installer name.
+
+    Args:
+        installer_name: installer name
+
+    Returns:
+        updated installer name
+    """
+    versions = {
+        "t/a": "trading as",
+        " ta ": " trading as ",
+        "a trading name of": "trading as",
+        "the trading name of": "trading as",
+    }
+    for item in versions:
+        installer_name = installer_name.replace(item, versions[item])
+
+    return installer_name
+
+
+def from_list_to_dictionary(list_values: list) -> dict:
+    """
+    Transforms a list into a dictionary where each key is matched to the first
+    value in the list. Creates a mapping between each item in the list and the first one.
+    Arg:
+        list_values: a list of values e.g. ["Company A", "Company B", "Company A trading as Company B"]
+    Returns:
+        Dictionary mapping all values to the first one
+        e.g. {"Company B":"Company A", "Company A trading as Company B": "Company A"}
+    """
+    return {value: list_values[0] for value in list_values[1:]}
+
+
+def map_position_of_subset_items(list_of_strings: list) -> dict:
+    """
+    Returns a mapping between the position of strings in a list which are a subset of each other.
+    Example:
+    set_of_strings = ["company a ltd", "a ltd", "a"]
+    Outputs {0: 2, 1: 2}
+    Note that, in the example above, 0 is initially mapped to 1 but then that's overwritten by 0 being mapped to 2.
+
+    Args:
+        list_of_strings: a list of strings
+    Returns:
+        A dictionary mapping the position of strings that are subset of another.
+    """
+    return {
+        i: j
+        for i in range(len(list_of_strings) - 1)
+        for j in range(i + 1, len(list_of_strings), 1)
+        if (list_of_strings[i] in list_of_strings[j])
+        or (list_of_strings[j] in list_of_strings[i])
+    }
+
+
+def position_to_value(list_of_strings: list, indices_dict: dict) -> dict:
+    """
+    Returns dictionary resulting from mapping a dictionary with indices as key-value pairs to a list of strings.
+    Args:
+        list_of_strings: list of strings, e.g. ["a", "b", "c"]
+        indices_dict: dictionary of indices, e.g. {0:2, 1:2}
+    Returns:
+        The mapped dictionary e.g. {"a":"c", "b":"c"}
+
+    """
+    return {
+        list_of_strings[key]: list_of_strings[indices_dict[key]]
+        for key in indices_dict.keys()
+    }
+
+
+def match_companies_house(company_name: str, api_key: str) -> pd.Series:
+    """
+    Fuzzy match between MCS company name and Companies House API company names and
+    returns Companies House information: company full address and company date of creation.
+
+    Args:
+        company_name: Company name used to query company house API
+        api_key: API key to request data from live company house API endpoint
+    Returns:
+        company_info: pandas Series with 2 values, full address and creation date
+    """
+
+    # E.g. "Flo Group LTD T/A Flo Renewables" -> "flo renewables"
+    company_name = company_name.lower()
+    if "trading as" in company_name:
+        company_name = deal_with_versions_of_trading_as(company_name).split(
+            " trading as "
+        )[1]
+
+    # endpoint url
+    base_url = f"https://api.company-information.service.gov.uk/search/companies?q={company_name}"
+
+    response = requests.get(base_url, auth=(api_key, ""))
+
+    # developer guidelines state you can make 600 requests per 5 mins
+    # this translates into 1 request every 0.5 seconds
+    time.sleep(0.5)
+
+    response_status_code = response.status_code
+
+    if response_status_code == 200:  # status code 200 means all good with request
+        company_data = response.json()
+        if (
+            (company_data is not None)
+            and ("items" in company_data.keys())
+            and len(company_data["items"]) > 0
+        ):
+            address_snippet = (
+                company_data["items"][0]["address_snippet"]
+                if ("address_snippet" in company_data["items"][0].keys())
+                else None
+            )
+            date_of_creation = (
+                company_data["items"][0]["date_of_creation"]
+                if "date_of_creation" in company_data["items"][0].keys()
+                else None
+            )
+
+            return pd.Series([address_snippet, date_of_creation])
+
+        else:
+            return pd.Series([None, None])
+    else:
+        if response_status_code >= 400 and response_status_code < 500:
+            raise Exception(
+                "Cannot get data, the program will stop!\nHTTP {}: {}".format(
+                    response_status_code, response.text
+                )
+            )
+        sleep_seconds = random.randint(5, 60)
+        print(
+            "Cannot get data, your program will sleep for {} seconds...\nHTTP {}: {}".format(
+                sleep_seconds, response_status_code, response.text
+            )
+        )
+        time.sleep(sleep_seconds)
+        return match_companies_house(company_name, api_key)
+
+
+# copied from process_mcs_utils.py and changed slightly
+def geocode_postcode(data: pd.DataFrame, geodata: pd.DataFrame) -> pd.DataFrame:
+    """
+    Updates data with latitude and longitude columns, by merging with geodata
+    on postode column (left merge, not to loose any data).
+    Also transforms postcode column by removing the space.
+
+    Args:
+        data: DataFrame with postcode column.
+        geodata: DataFrame with postcode, latitude and longitude columns.
+    """
+
+    geodata["postcode"] = geodata["postcode"].str.replace(" ", "")
+
+    data["postcode"] = data["postcode"].str.upper().str.replace(" ", "")
+    data = data.merge(
+        geodata[["postcode", "latitude", "longitude"]], how="left", on="postcode"
+    )
+
+    return data
+
+
 def clean_concat_installers(data):
-    """Cleans concatenated installers data by:
+    """
+    Legacy function: only used for quarterly installers data,
+    not for historical installers.
+
+    Cleans concatenated installers data by:
         - dropping duplicate company names;
         - removing columns with identical values;
         - merging similar columns.
@@ -152,80 +341,6 @@ def clean_concat_installers(data):
     ]
 
     return data.reset_index(drop=True)
-
-
-def geocode_postcode(data, geodata):
-    """geocodes dataset with 'postcode' column.
-    Args:
-        data (pd.DataFrame): DataFrame with 'postcode' column.
-        geodata (pd.DataFrame): DataFrame with 'postcode', 'latitude' and 'longitude' columns.
-    Returns:
-        data_with_geodata (pd.DataFrame): Merged data DataFrame with 'latitude' and 'longitude'.
-    """
-
-    geodata["postcode"] = geodata["postcode"].str.replace(" ", "")
-
-    if "postcode" in data.columns:
-        data["postcode"] = data["postcode"].str.upper().str.replace(" ", "")
-        data_with_postcode = pd.merge(
-            data, geodata[["postcode", "latitude", "longitude"]], on="postcode"
-        )
-        print(
-            f"lost {len(list(set(data.postcode) - set(geodata.postcode)))} postcodes in merge."
-        )
-        return data_with_postcode
-    else:
-        print("'postcode' not found in data columns.")
-
-
-def match_companies_house(company_name, api_key):
-    """Fuzzy matches company name with company house API and
-        returns companies house information.
-    Args:
-        company_name (str): Company name to query company house API with.
-        api_key (str): Api key to call live company house API end point.
-    Returns:
-        company_info (dict): Dictionary of company house data associated to company name.
-    """
-
-    if len(company_name.split()) == 1:
-        company_name
-    else:
-        company_name = "+".join(company_name.split())
-
-    company_info = dict()
-    base_url = "https://api.company-information.service.gov.uk/search/companies?q={company_name}".format(
-        company_name=company_name
-    )
-    response = requests.get(base_url, auth=(api_key, ""))
-
-    if response.status_code == 200:
-        company_data = response.json()
-        if company_data is not None:
-            if "items" in company_data.keys():
-                if company_data["items"] != []:
-                    closest_match = company_data["items"][
-                        0
-                    ]  # closest match according to companies house API
-
-                    company_info[company_name.replace("+", " ")] = {
-                        "company_number": closest_match["company_number"],
-                        "title": closest_match["title"],
-                        "company_status": closest_match["company_status"],
-                        "description_identifier": closest_match[
-                            "description_identifier"
-                        ],
-                        "address_snippet": closest_match["address_snippet"],
-                    }
-        return company_info
-
-    elif response.status_code == 429:
-        print(f"{response.status_code} error! Sleeping for 5 mins...")
-        time.sleep(
-            301
-        )  # sleep for 5 minutes as developer guidelines state you can make 600 requests per 5 mins
-    else:
-        print(f"{response.status_code} error!")
 
 
 def remove_punctuation(address):
@@ -289,19 +404,100 @@ def extract_token_set(address, postcode, max_token_length):
     return valid_token_set
 
 
-# %%
-
-if __name__ == "__main__":
-    installer_company_data_path = base_config.MCS_RAW_INSTALLER_CONCAT_S3_PATH
-    uk_geo_path = base_config.POSTCODE_TO_COORD_PATH
-    cleaned_installations_path = base_config.PREPROC_GEO_MCS_INSTALLATIONS_PATH
-    cleaned_installer_company_path = base_config.PREPROC_MCS_INSTALLER_COMPANY_PATH
-    bucket_name = base_config.BUCKET_NAME
-
-    installer_company_data = load_s3_data(bucket_name, installer_company_data_path)
-    uk_geo_data = load_s3_data(bucket_name, uk_geo_path)
-
-    ## preprocess different columns
-    installer_company_data = clean_concat_installers(
-        installer_company_data
-    ).reset_index(drop=True)
+## ----- Legacy functions below -----
+# The functions below were originally created by India and used to process the installers data we were receving
+# from MCS each quarter. Now that we are working with the historical table of installers, we had to change these
+# functions slighly as per above. We kept the original ones for reference, if we ever need to go back to them to
+# repeat any analysis.
+# def clean_company_name(company_name):
+#    """cleans installation company name by:
+#        - removing "company" stopwords like ltd, limited;
+#        - lowers company name
+#        - removes punctuation
+#    Args:
+#        company_name (str): Name of company to clean.
+#    Returns:
+#        company_name (str): Cleaned company name.
+#    """
+#    company_stopwords = ["ltd", "ltd.", "limited", "limited.", "old", "account", "t/a"]
+#
+#    company_name = company_name.lower().translate(str.maketrans("", "", "()-"))
+#    company_name = [
+#        comp for comp in company_name.split() if comp not in company_stopwords
+#    ]
+#
+#    return " ".join(company_name)
+#
+#
+# def geocode_postcode(data, geodata):
+#    """geocodes dataset with 'postcode' column.
+#    Args:
+#        data (pd.DataFrame): DataFrame with 'postcode' column.
+#        geodata (pd.DataFrame): DataFrame with 'postcode', 'latitude' and 'longitude' columns.
+#    Returns:
+#        data_with_geodata (pd.DataFrame): Merged data DataFrame with 'latitude' and 'longitude'.
+#    """
+#
+#    geodata["postcode"] = geodata["postcode"].str.replace(" ", "")
+#
+#    if "postcode" in data.columns:
+#        data["postcode"] = data["postcode"].str.upper().str.replace(" ", "")
+#        data_with_postcode = pd.merge(
+#            data, geodata[["postcode", "latitude", "longitude"]], on="postcode"
+#        )
+#        print(
+#            f"lost {len(list(set(data.postcode) - set(geodata.postcode)))} postcodes in merge."
+#        )
+#        return data_with_postcode
+#    else:
+#        print("'postcode' not found in data columns.")
+#
+#
+# def match_companies_house(company_name, api_key):
+#    """Fuzzy matches company name with company house API and
+#        returns companies house information.
+#    Args:
+#        company_name (str): Company name to query company house API with.
+#        api_key (str): Api key to call live company house API end point.
+#    Returns:
+#        company_info (dict): Dictionary of company house data associated to company name.
+#    """
+#
+#    if len(company_name.split()) == 1:
+#        company_name
+#    else:
+#        company_name = "+".join(company_name.split())
+#
+#    company_info = dict()
+#    base_url = "https://api.company-information.service.gov.uk/search/companies?q={company_name}".format(
+#        company_name=company_name
+#    )
+#    response = requests.get(base_url, auth=(api_key, ""))
+#
+#    if response.status_code == 200:
+#        company_data = response.json()
+#        if company_data is not None:
+#            if "items" in company_data.keys():
+#                if company_data["items"] != []:
+#                    closest_match = company_data["items"][
+#                        0
+#                    ]  # closest match according to companies house API
+#
+#                    company_info[company_name.replace("+", " ")] = {
+#                        "company_number": closest_match["company_number"],
+#                        "title": closest_match["title"],
+#                        "company_status": closest_match["company_status"],
+#                        "description_identifier": closest_match[
+#                            "description_identifier"
+#                        ],
+#                        "address_snippet": closest_match["address_snippet"],
+#                    }
+#        return company_info
+#
+#    elif response.status_code == 429:
+#        print(f"{response.status_code} error! Sleeping for 5 mins...")
+#        time.sleep(
+#            301
+#        )  # sleep for 5 minutes as developer guidelines state you can make 600 requests per 5 mins
+#    else:
+#        print(f"{response.status_code} error!")
